1st attempt at reading a file
When running this code, an error appears stating "usb.core.USBError: [Errno 13] Access denied (insufficient permissions)". My hypothesis is that we are unable to set the device to default configurations. I attempted to fix this by commenting out line 21, however, this only created an endless loop and stopped the program from functioning.